refactor: added auth hooks for handling sign and signup logics
