+++ apps/web/src/app/(auth)/login/page.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { signIn } from "@opencut/auth/client";
 import { Button } from "@/components/ui/button";
 import {
   Card,
@@ -10,7 +9,7 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import { Suspense, useState } from "react";
+import { Suspense } from "react";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Separator } from "@/components/ui/separator";
@@ -18,121 +17,22 @@ import Link from "next/link";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { ArrowLeft, Loader2 } from "lucide-react";
 import { GoogleIcon } from "@/components/icons";
-
-function LoginForm() {
-  const router = useRouter();
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState<string | null>(null);
-  const [isEmailLoading, setIsEmailLoading] = useState(false);
-  const [isGoogleLoading, setIsGoogleLoading] = useState(false);
-
-  const handleLogin = async () => {
-    setError(null);
-    setIsEmailLoading(true);
-
-    const { error } = await signIn.email({
-      email,
-      password,
-    });
-
-    if (error) {
-      setError(error.message || "An unexpected error occurred.");
-      setIsEmailLoading(false);
-      return;
-    }
-
-    router.push("/editor");
-  };
-
-  const handleGoogleLogin = async () => {
-    setError(null);
-    setIsGoogleLoading(true);
-
-    try {
-      await signIn.social({
-        provider: "google",
-        callbackURL: "/editor",
-      });
-    } catch (error) {
-      setError("Failed to sign in with Google. Please try again.");
-      setIsGoogleLoading(false);
-    }
-  };
-
-  const isAnyLoading = isEmailLoading || isGoogleLoading;
-
-  return (
-    <div className="flex flex-col space-y-6">
-      {error && (
-        <Alert variant="destructive">
-          <AlertTitle>Error</AlertTitle>
-          <AlertDescription>{error}</AlertDescription>
-        </Alert>
-      )}
-
-      <Button
-        onClick={handleGoogleLogin}
-        variant="outline"
-        size="lg"
-        disabled={isAnyLoading}
-      >
-        {isGoogleLoading ? (
-          <Loader2 className="animate-spin" />
-        ) : (
-          <GoogleIcon />
-        )}{" "}
-        Continue with Google
-      </Button>
-      <div className="relative">
-        <div className="absolute inset-0 flex items-center">
-          <Separator className="w-full" />
-        </div>
-        <div className="relative flex justify-center text-xs uppercase">
-          <span className="bg-background px-2 text-muted-foreground">
-            Or continue with
-          </span>
-        </div>
-      </div>
-      <div className="space-y-4">
-        <div className="space-y-2">
-          <Label htmlFor="email">Email</Label>
-          <Input
-            id="email"
-            type="email"
-            placeholder="m@example.com"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            disabled={isAnyLoading}
-            className="h-11"
-          />
-        </div>
-        <div className="space-y-2">
-          <Label htmlFor="password">Password</Label>
-          <Input
-            id="password"
-            type="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            disabled={isAnyLoading}
-            className="h-11"
-          />
-        </div>
-        <Button
-          onClick={handleLogin}
-          disabled={isAnyLoading || !email || !password}
-          className="w-full h-11"
-          size="lg"
-        >
-          {isEmailLoading ? <Loader2 className="animate-spin" /> : "Sign in"}
-        </Button>
-      </div>
-    </div>
-  );
-}
+import { useLogin } from "@/hooks/auth/useLogin";
 
 export default function LoginPage() {
   const router = useRouter();
+  const {
+    email,
+    setEmail,
+    password,
+    setPassword,
+    error,
+    isAnyLoading,
+    isEmailLoading,
+    isGoogleLoading,
+    handleLogin,
+    handleGoogleLogin,
+  } = useLogin();
 
   return (
     <div className="flex h-screen items-center justify-center relative">
@@ -158,17 +58,81 @@ export default function LoginPage() {
               </div>
             }
           >
-            <LoginForm />
+            <div className="flex flex-col space-y-6">
+              {error && (
+                <Alert variant="destructive">
+                  <AlertTitle>Error</AlertTitle>
+                  <AlertDescription>{error}</AlertDescription>
+                </Alert>
+              )}
+
+              <Button
+                onClick={handleGoogleLogin}
+                variant="outline"
+                size="lg"
+                disabled={isAnyLoading}
+              >
+                {isGoogleLoading ? (
+                  <Loader2 className="animate-spin" />
+                ) : (
+                  <GoogleIcon />
+                )}{" "}
+                Continue with Google
+              </Button>
+              <div className="relative">
+                <div className="absolute inset-0 flex items-center">
+                  <Separator className="w-full" />
+                </div>
+                <div className="relative flex justify-center text-xs uppercase">
+                  <span className="bg-background px-2 text-muted-foreground">
+                    Or continue with
+                  </span>
+                </div>
+              </div>
+              <div className="space-y-4">
+                <div className="space-y-2">
+                  <Label htmlFor="email">Email</Label>
+                  <Input
+                    id="email"
+                    type="email"
+                    placeholder="m@example.com"
+                    value={email}
+                    onChange={(e) => setEmail(e.target.value)}
+                    disabled={isAnyLoading}
+                    className="h-11"
+                  />
+                </div>
+                <div className="space-y-2">
+                  <Label htmlFor="password">Password</Label>
+                  <Input
+                    id="password"
+                    type="password"
+                    value={password}
+                    onChange={(e) => setPassword(e.target.value)}
+                    disabled={isAnyLoading}
+                    className="h-11"
+                  />
+                </div>
+                <Button
+                  onClick={handleLogin}
+                  disabled={isAnyLoading || !email || !password}
+                  className="w-full h-11"
+                  size="lg"
+                >
+                  {isEmailLoading ? <Loader2 className="animate-spin" /> : "Sign in"}
+                </Button>
+              </div>
+            </div>
+            <div className="mt-6 text-center text-sm">
+              Don't have an account?{" "}
+              <Link
+                href="/signup"
+                className="font-medium text-primary underline-offset-4 hover:underline"
+              >
+                Sign up
+              </Link>
+            </div>
           </Suspense>
-          <div className="mt-6 text-center text-sm">
-            Don't have an account?{" "}
-            <Link
-              href="/signup"
-              className="font-medium text-primary underline-offset-4 hover:underline"
-            >
-              Sign up
-            </Link>
-          </div>
         </CardContent>
       </Card>
     </div>
